fix: add TurboModule specs and update iOS implementation for new architecture compatibility

+++ ios/RNSensorsGyroscope.h
@@ -22,7 +22,7 @@
          rejecter:(RCTPromiseRejectBlock) reject;
 - (void) setUpdateInterval:(double) interval;
 - (void) getUpdateInterval:(RCTResponseSenderBlock) cb;
-- (void) setLogLevel:(int) level;
+- (void) setLogLevel:(double) level;
 - (void) getData:(RCTResponseSenderBlock) cb;
 - (void) startUpdates;
 - (void) stopUpdates;
